+++ R/predict.R
@@ -12,6 +12,13 @@
 #'   fitted model object are used.
 #' @param se.fit A logical indicating if standard errors are returned.
 #'   The default is \code{FALSE}.
+#' @param scale A numeric constant by which to scale the regular standard errors and intervals.
+#'   Similar to but slightly different than \code{scale} for [stats::predict.lm()], because
+#'   predictions form a spatial model may have different residual variances for each
+#'   observation in \code{newdata}. The default is \code{NULL}, which returns
+#'   the regular standard errors and intervals.
+#' @param df Degrees of freedom to use for confidence or prediction intervals
+#'   (ignored if \code{scale} is not specified). The default is \code{Inf}.
 #' @param interval Type of interval calculation. The default is \code{"none"}.
 #'   Other options are \code{"confidence"} (for confidence intervals) and
 #'   \code{"prediction"} (for prediction intervals).
@@ -106,7 +113,7 @@
 #' predict(spmod, sulfate_preds)
 #' predict(spmod, sulfate_preds, interval = "prediction")
 #' augment(spmod, newdata = sulfate_preds, interval = "prediction")
-predict.splm <- function(object, newdata, se.fit = FALSE, interval = c("none", "confidence", "prediction"),
+predict.splm <- function(object, newdata, se.fit = FALSE, scale = NULL, df = Inf, interval = c("none", "confidence", "prediction"),
                          level = 0.95, type = c("response", "terms"), local, terms = NULL, ...) {
 
   # match interval argument so the three display
@@ -116,6 +123,11 @@ predict.splm <- function(object, newdata, se.fit = FALSE, interval = c("none", "
   # deal with local
   if (missing(local)) {
     local <- NULL
+  }
+
+  # check scale is numeric (if specified)
+  if (!is.null(scale) && !is.numeric(scale)) {
+    stop("scale must be numeric.", call. = FALSE)
   }
 
   # error if newdata missing from arguments and object
@@ -218,7 +230,7 @@ predict.splm <- function(object, newdata, se.fit = FALSE, interval = c("none", "
 
   # call terms if needed
   if (type == "terms") {
-    return(predict_terms(object, newdata_model, se.fit, interval, level, add_newdata_rows, terms, ...))
+    return(predict_terms(object, newdata_model, se.fit, scale, df, interval, level, add_newdata_rows, terms, ...))
   }
 
   # storing newdata as a list
@@ -429,6 +441,9 @@ predict.splm <- function(object, newdata, se.fit = FALSE, interval = c("none", "
       if (se.fit) {
         vars <- vapply(pred_splm, function(x) x$var, numeric(1))
         se <- sqrt(vars)
+        if (!is.null(scale)) {
+          se <- se * scale
+        }
         if (add_newdata_rows) {
           names(fit) <- object$missing_index
           names(se) <- object$missing_index
@@ -450,8 +465,15 @@ predict.splm <- function(object, newdata, se.fit = FALSE, interval = c("none", "
       }
       vars <- vapply(pred_splm, function(x) x$var, numeric(1))
       se <- sqrt(vars)
+      if (!is.null(scale)) {
+        se <- se * scale
+        df <- df
+      } else {
+        df <- Inf
+      }
+      tstar <- qt(1 - (1 - level) / 2, df = df)
       # tstar <- qt(1 - (1 - level) / 2, df = object$n - object$p)
-      tstar <- qnorm(1 - (1 - level) / 2)
+      # tstar <- qnorm(1 - (1 - level) / 2)
       lwr <- fit - tstar * se
       upr <- fit + tstar * se
       fit <- cbind(fit, lwr, upr)
@@ -479,8 +501,13 @@ predict.splm <- function(object, newdata, se.fit = FALSE, interval = c("none", "
     newdata_model_list <- split(newdata_model, seq_len(NROW(newdata_model)))
     vars <- as.numeric(vapply(newdata_model_list, function(x) crossprod(x, vcov(object) %*% x), numeric(1)))
     se <- sqrt(vars)
-    # tstar <- qt(1 - (1 - level) / 2, df = object$n - object$p)
-    tstar <- qnorm(1 - (1 - level) / 2)
+    if (!is.null(scale)) {
+      se <- se * scale
+      df <- df
+    } else {
+      df <- Inf
+    }
+    tstar <- qt(1 - (1 - level) / 2, df = df)
     lwr <- fit - tstar * se
     upr <- fit + tstar * se
     fit <- cbind(fit, lwr, upr)
@@ -506,7 +533,7 @@ predict.splm <- function(object, newdata, se.fit = FALSE, interval = c("none", "
 #' @method predict spautor
 #' @order 2
 #' @export
-predict.spautor <- function(object, newdata, se.fit = FALSE, interval = c("none", "confidence", "prediction"),
+predict.spautor <- function(object, newdata, se.fit = FALSE, scale = NULL, df = Inf, interval = c("none", "confidence", "prediction"),
                             level = 0.95, type = c("response", "terms"), local, terms = NULL, ...) {
 
   # match interval argument so the three display
@@ -518,16 +545,19 @@ predict.spautor <- function(object, newdata, se.fit = FALSE, interval = c("none"
     local <- NULL
   }
 
-  # error if newdata missing from arguments and object
-  if (missing(newdata) && is.null(object$newdata)) {
-    stop("No missing data to predict. newdata must be specified in the newdata argument or object$newdata must be non-NULL.", call. = FALSE)
-  }
-
-
-
   # deal with local
   if (is.null(local)) {
     local <- FALSE
+  }
+
+  # check scale is numeric (if specified)
+  if (!is.null(scale) && !is.numeric(scale)) {
+    stop("scale must be numeric.", call. = FALSE)
+  }
+
+  # error if newdata missing from arguments and object
+  if (missing(newdata) && is.null(object$newdata)) {
+    stop("No missing data to predict. newdata must be specified in the newdata argument or object$newdata must be non-NULL.", call. = FALSE)
   }
 
   # write newdata if predicting missing data
@@ -572,7 +602,7 @@ predict.spautor <- function(object, newdata, se.fit = FALSE, interval = c("none"
   # call terms if needed
   if (type == "terms") {
     # may want to add add_newdata_rows if we allow spautor prediction for the observed model matrix
-    return(predict_terms(object, newdata_model, se.fit, interval, level, add_newdata_rows = TRUE, terms, ...))
+    return(predict_terms(object, newdata_model, se.fit, scale, df, interval, level, add_newdata_rows = TRUE, terms, ...))
   }
 
 
@@ -664,6 +694,9 @@ predict.spautor <- function(object, newdata, se.fit = FALSE, interval = c("none"
       if (se.fit) {
         vars <- vapply(pred_spautor, function(x) x$var, numeric(1))
         se <- sqrt(vars)
+        if (!is.null(scale)) {
+          se <- se * scale
+        }
         names(fit) <- object$missing_index
         names(se) <- object$missing_index
         return(list(fit = fit, se.fit = se))
@@ -681,8 +714,13 @@ predict.spautor <- function(object, newdata, se.fit = FALSE, interval = c("none"
       }
       vars <- vapply(pred_spautor, function(x) x$var, numeric(1))
       se <- sqrt(vars)
-      # tstar <- qt(1 - (1 - level) / 2, df = object$n - object$p)
-      tstar <- qnorm(1 - (1 - level) / 2)
+      if (!is.null(scale)) {
+        se <- se * scale
+        df <- df
+      } else {
+        df <- Inf
+      }
+      tstar <- qt(1 - (1 - level) / 2, df = df)
       lwr <- fit - tstar * se
       upr <- fit + tstar * se
       fit <- cbind(fit, lwr, upr)
@@ -705,8 +743,13 @@ predict.spautor <- function(object, newdata, se.fit = FALSE, interval = c("none"
     }
     vars <- as.numeric(vapply(newdata_model_list, function(x) crossprod(x, vcov(object) %*% x), numeric(1)))
     se <- sqrt(vars)
-    # tstar <- qt(1 - (1 - level) / 2, df = object$n - object$p)
-    tstar <- qnorm(1 - (1 - level) / 2)
+    if (!is.null(scale)) {
+      se <- se * scale
+      df <- df
+    } else {
+      df <- Inf
+    }
+    tstar <- qt(1 - (1 - level) / 2, df = df)
     lwr <- fit - tstar * se
     upr <- fit + tstar * se
     fit <- cbind(fit, lwr, upr)