+++ NEWS.md
@@ -12,6 +12,7 @@
 * Updated `logLik()` to match `lm()` and `glm()` behavior. `logLik()` now returns a vector with class `logLik` and attributes `nobs` and `df`.
 * Added support for using `AIC()` and `BIC()` from `stats` and removed `spmodel`-specific `AIC()` and `BIC()` methods.
 * Added support for `"terms"` prediction for `splm()`, `spautor()`, `spglm()`, and `spgautor()` models.
+* Added `scale` and `df` arguments to `predict()` for `splm()` and `spautor()` models.
 
 
 # spmodel 0.7.0